refactor: moving routes from mongo connection file to a separate one
