+++ backend/server/routers/index.js
@@ -2,4 +2,5 @@ module.exports = function(app, server){
     require('./public')(app);    
     require('./health')(app);
     require('./mongo')(app, server);
+    require('./routes')(app);
 };
